--- core/src/main/kotlin/com/tdt4240/group3/config/unit/MovementRules.kt
@@ -1,6 +1,0 @@
-package com.tdt4240.group3.config.unit
-
-data class MovementRules(
-    val moveRange: Int = 2,
-    val canCrossWater: Boolean = false
-)
